airflow-docker/requirements.txt
pandas==2.2.0
isodate==0.7.2
apache-airflow==3.0.1
python-dotenv==0.21.0
